requirements.txt
coverage==3.7.1
nose==1.3.4
python-magic==0.4.6
requests==2.4.3
typedecorator==0.0.4
